requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
asttokens==2.1.0
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
colorama==0.4.6
contourpy==1.0.6
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
disnake==2.7.0
entrypoints==0.4
executing==1.2.0
fonttools==4.38.0
frozenlist==1.3.3
idna==3.4
ipykernel==6.17.1
ipython==8.6.0
jedi==0.18.2
jupyter_client==7.4.7
jupyter_core==5.0.0
kiwisolver==1.4.4
llvmlite==0.38.0
lxml==4.9.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
multidict==6.0.2
mysql-connector-python==8.0.31
mysqlclient==2.1.1
nest-asyncio==1.5.6
numpy==1.23.4
packaging==21.3
pandas==1.5.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.5.4
prompt-toolkit==3.0.33
protobuf==3.20.1
psutil==5.9.4
pure-eval==0.2.2
pycosat==0.6.3
pycurl==7.44.1
Pygments==2.13.0
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pyreadline==2.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
pywin32==305
pyzmq==24.0.1
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
stack-data==0.6.1
TBB==0.2
tornado==6.2
traitlets==5.5.0
urllib3==1.26.12
wcwidth==0.2.5
wincertstore==0.2
xmltodict==0.13.0
yarl==1.8.1
